+++ source/digioz.Forum/digioz.Forum/Models/ForumBookmark.cs
@@ -5,7 +5,7 @@ namespace digioz.Forum.Models;
 
 public partial class ForumBookmark
 {
-    public int TopicId { get; set; }
+    public long TopicId { get; set; }
 
     public int UserId { get; set; }
 }
