+++ source/digioz.Forum/digioz.Forum/Models/ForumLog.cs
@@ -5,7 +5,7 @@ namespace digioz.Forum.Models;
 
 public partial class ForumLog
 {
-    public int LogId { get; set; }
+    public long LogId { get; set; }
 
     public byte LogType { get; set; }
 
@@ -13,9 +13,9 @@ public partial class ForumLog
 
     public long ForumId { get; set; }
 
-    public int TopicId { get; set; }
+    public long TopicId { get; set; }
 
-    public int PostId { get; set; }
+    public long PostId { get; set; }
 
     public int ReporteeId { get; set; }
